+++ examples/drag_and_drop.rs
@@ -10,6 +10,7 @@ fn main() {
         DefaultPickingPlugins,
         SpinPlugin,
     ))
+    .insert_resource(DebugPickingMode::Normal)
     .add_systems(Startup, setup);
     #[cfg(feature = "backend_egui")]
     app.add_plugins(bevy_egui::EguiPlugin);
@@ -29,7 +30,7 @@ fn setup(
         let z = 0.5 + x as f32 * 0.1;
         commands.spawn((
             MaterialMesh2dBundle {
-                mesh: meshes.add(Mesh::from(shape::Quad::default())).into(),
+                mesh: meshes.add(Rectangle::default()).into(),
                 transform: Transform::from_xyz(x as f32 * 200.0, 0.0, z)
                     .with_scale(Vec3::splat(100.)),
                 material: materials.add(ColorMaterial::from(Color::hsl(0.0, 1.0, z))),