+++ examples/render_layers.rs
@@ -1,6 +1,6 @@
 //! Demonstrates that picking respects render layers and order.
 
-use bevy::{core_pipeline::clear_color::ClearColorConfig, prelude::*};
+use bevy::{prelude::*, render::camera::ClearColorConfig};
 use bevy_mod_picking::prelude::*;
 use bevy_render::view::RenderLayers;
 
@@ -10,6 +10,7 @@ fn main() {
             DefaultPlugins.set(low_latency_window_plugin()),
             DefaultPickingPlugins,
         ))
+        .insert_resource(DebugPickingMode::Normal)
         .add_systems(Startup, setup)
         .run();
 }
@@ -23,8 +24,11 @@ fn setup(
     // plane
     commands.spawn((
         PbrBundle {
-            mesh: meshes.add(shape::Plane::from_size(5.0).into()),
-            material: materials.add(Color::rgb(0.3, 0.5, 0.3).into()),
+            mesh: meshes.add(bevy_render::mesh::PlaneMeshBuilder {
+                half_size: Vec2::splat(2.5),
+                ..default()
+            }),
+            material: materials.add(Color::rgb(0.3, 0.5, 0.3)),
             ..default()
         },
         PickableBundle::default(), // <- Makes the mesh pickable.
@@ -32,14 +36,8 @@ fn setup(
     // sphere
     commands.spawn((
         PbrBundle {
-            mesh: meshes.add(
-                Mesh::try_from(shape::Icosphere {
-                    subdivisions: 3,
-                    radius: 0.5,
-                })
-                .unwrap(),
-            ),
-            material: materials.add(Color::rgb(0.8, 0.7, 0.6).into()),
+            mesh: meshes.add(Sphere::default()),
+            material: materials.add(Color::rgb(0.8, 0.7, 0.6)),
             transform: Transform::from_xyz(0.0, 0.0, 0.0),
             ..default()
         },
@@ -49,7 +47,6 @@ fn setup(
     // light
     commands.spawn(PointLightBundle {
         point_light: PointLight {
-            intensity: 1500.0,
             shadows_enabled: true,
             ..default()
         },
@@ -67,11 +64,8 @@ fn setup(
     commands.spawn((
         Camera3dBundle {
             transform: camera_transform,
-            camera_3d: Camera3d {
-                clear_color: ClearColorConfig::None,
-                ..default()
-            },
             camera: Camera {
+                clear_color: ClearColorConfig::None,
                 // renders after / on top of the main camera
                 order: 1,
                 ..default()